had to copy in Simp....
I want to be able to measure the text without calling prepare_text(). The problem is that a lot of the font stuff in Simp is behind a #scope_file in a way that makes it impossible to just measure the glyphs myself. Hopefully I can get this changed in the official module somehow...